Part 1
In this part, my team and I used Python libraries and code to understand and analyse a dataset to identify common trends, highlight 
the data quality issues and correct those problems to enhance the data quality for the future data 
mining task. We aimed to prepared dataset D1 to build data mining models in the part 2.